fix: connectorTemplateId state
We opted to use the `??` operator when defaulting the
`connectorTempalateId` property with the intention if it wasn't _set_
then it should default to `swagger-connector-template`. This created an
issue with setting the default React state of `connectorTempalateId` to
`""` as the `??` operator would use that value instead of the dafault
`swagger-connector-template`. That is, before when we used the ternary
operator (`?:`) it would consider the empty string as falsy and it would
use the default value instead.
This changes the default React state to `undefined` so that the logic
using the `??` operator branches to using the correct default value.

Ref. https://issues.redhat.com/browse/ENTESB-17811