chore: align versions in integration and S2I
By importing `spring-boot-dependencies` BOM we have imported
dependencies with version from there, so any alignment we did in the
Syndesis as a Maven project, i.e. making sure that a connector builds
with a certain version of Netty was not applied to the generated
integration project. So we would end up with a connector build with a
version of a dependency that it would not run with.

This adds the BOMs of projects that are also connector dependencies in
the integration BOM, copying the versions from the Syndesis Maven
project and changes the template for generating the integration POM not
to include `spring-boot-dependencies`. The end result is that if we keep
updating the versions of dependencies in the integration BOM the same
versions we compile and test with will be used at runtime in the
integration.