fix: s2i dependency download
The `go-offline-maven-plugin` was invoked without specifying
`-DfailOnErrors=true`, so it could fail silently. The plugin also runs
against the parent POM, the modules are skipped, and the repository
configuration is read from the project, the provided settings file helps
with caching but if a repository is not defined in the parent POM it
will not be used.

These two things combined lead to the S2I image missing dependencies
that are not available on Maven central.

Since the plugin runs only for the parent POM it doesn't make much sense
to run it in parallel, so that was removed.

In the project generator application, the application used to generate
projects for dependency gathering the three repositories that we know
are required are set in addition to any repositories that could be
configured in the Spring configuration environment.