chore: don't use custom POM template
The `pom.xml.mustache` template from the project generator can easily
get out of sync with this template customized in the `test-support`. We
can end up testing against a project generated with a different
configuration than a project that will be run. We want the integration
tests to be as close as possible to the production workloads.