refactor: mongodb connector tests
These are some of the slowest and flakiest tests we have so refactoring
them would benefit us going forward.
This converts some of the tests from integration tests to unit tests and
uses a single MongoDB instance for running the integration tests. Also
the tests are now run in parallel to gain performance.