fix: force newer Gson version
The Apache Olingo 2 depends on a version of Gson (2.4) that is too old
for the Spring Boot Gson starter which requires 2.6 or newer. This
change forces the version to the version brought in by Jaeger (2.8.7),
and it's the same version that was brought in by the now removed
`spring-boot-dependencies` BOM.

Ref. https://issues.redhat.com/browse/ENTESB-17885