chore: remove server-builder-maven-plugin
I was looking into an issue reported by Snyk and traced it to a
dependency declared in `server-builder-maven-plugin`, I've also found
that the dependency is not used (not sure why the dependency checks
didn't fail), but also that the whole plugin is not used.

This plugin has two MOJOs:
 - extract-connector-descriptors, that is not used
 - generate-mapper-inspections, that generates AtlasMap inspection data
   for any Connnector in `deployment.json` that contains data shapes; we
   don't have any Connectors in `deployment.json` that contain data
   shapes

This removes the plugin entirely.