fix(server): correct totalCount in ListResult
When filters (e.g. pagination) are used the `totalCount` gets reset to
whatever the filter outcome size is (e.g. page size), and the initial
`totalCount` that reflects the number of available items is lost.

This refactors the `ListResult` to clearly support two cases: partial or
complete item set was provided; also for the existing filtering
operations partial case is used with the `totalCount` of the original
item set.