chore: enforce and converge dependencies
This enforces the dependency convergence via Maven Enforcer rule and
converges the dependencies to a single version. In cases this meant
removing a Syndesis-declared dependency, upgrading or (rarely)
downgrading a dependency.

Going forward this will fail the build in case non-converged
dependencies are found and hopefully solve dependency issues as they're
introduced.