refactor: use embedded resources
The way we bundle resources within the executable is problematic.
Resources need to be included by running code generation, so if skipped
older version of the resource will be included; and the generated .go
files contain the resources as byte arrays in hex, nullifying the
utility of diff tools and making resolving conflicts difficult.

This instead uses the `//go:embed` directive to embed the resources at
compilation time, making sure that the most recent version of the
resource is included and making potential merge conflicts explicit and
simpler to resolve.