refactor(ui): replace Learna with Lage
Lage[1] makes a good replacement for Lerna because of strict topological
workflow -- so tasks in workspaces are run in accordance to the how they
depend on each other; and because of the incremental cache -- so tasks
are skipped if inputs have not changed.

This dramatically decreases the build time by allowing the `build` task
to be run in parallel, skipping over tasks for which the cache is hot.

[1] https://microsoft.github.io/lage/