feat(ui): add a warning when mapped shapes differ
If a input data shape of a subsequent step or the output of a previous
step, relative to the data mapper step has changed after the mapping was
configured a warning is shown inviting the end user to edit the mapping
in order to apply any changes.

This is done by tracking the input or output data shapes for changes and
setting the `inputUpdatedAt` or `outputUpdatedAt` metadata properties to
the timestamp of when they were changed. This is in turn compared to the
`updatedAt` timestamp on the mapping step to see if the mapping step is
outdated.