fix(ui): update mapping update timestamp on insert
When the mapping step is added to the flow we also need to set it's
updated timestamp. There is a case when previous or next step has
updated timestamp set to `undefined`, so compared to the newly inserted
mapping step it is up to date even though the mapping step might need
to be updated because of a data shape change on that step. This could
occur, for example, when a subsequent step's action is replaced with
an action with a different data shape after a newly added mapping step,
both would have their respective timestamps set to `undefined`.