fix(ui): mapping step recommendation refinement
If we change the input data shape of a step after the mapper, don't
recommend that a mapping step is added when updating the existing
mapping step is a better solution. An additional condition was added to
the existing condition determining if the mapping step should be added
to check if previous step is a mapping step and if that step is outdated
relative to the current step.