refactor(ui): mapper update warning
There are four cases when the mapping step needs updating:
 * one of the source steps has changed it's output
 * the target step's input is not compatible with the mapper's output
 * one of the source steps has gone missing
 * the target step has gone missing

This change accounts for those. Since we're no longer looking at
timestamps for input data shapes, we no longer need to set those so that
bit was removed. Tests for, and some combinations, the above cases are
provided.

To make further maintenance easier _Input_ and _Output_ qualifiers were
added to functions.