fix(server): metrics with ubi/openjdk base image
The Prometheus JMX exporter is configured with a different set of
environment variables than in the fabric8/java image that we used to
use. This switches to using those and also makes sure that the
exporter's configuration file is placed in `/deployments/data` instead
of `/tmp/src` which gets deleted by the S2I build.