refactor(s2i): less permission changes
We switch between UID `0` and `1000` and finally settle on `185/jboss`.
We don't need to do that at all if we start as UID `jboss` and make sure
all steps are performed as `jboss` user.
This makes it easier to maintain and less gymnastics is required to make
sure that the S2I build succeeds.
As noted on #8451, the `ADD --chown=...` syntax is not supported on
OpenShift 3.11.