refactor(server): simplify unpublishing
This refactors the unpublishing step to perform a much simpler operation
of scaling to zero the existing deployment found via deployment name.

Also refactors the placement of the scale annotation to the
`DeploymentConfig`'s metadata rather than on container's metadata. The
whole deployment is scaled not an individual container. This makes for a
simpler update that doesn't require a new revision of the deployment and
prevents further re-deployments.

Subsequent commit will also include this change at deployment time.