fix(extension): compile issues Spring Boot upgrade
The `AbstractDependencyFilterMojo:filterDependencies` method that we
override has become `final`, so this copies and repackages
`AbstractDependencyFilterMojo` and all of the code that is referenced
from it in a package private/protected way so we can compile the plugin.

The Spring Boot code included here is ASF licensed, so it should be
okay.

Some code quality/style checks changes were required.