fix(api-generator): don't load remote schemas
We should not reach out and load remote XML schemas from
schemas.xmlsoap.org.

Background: it seems that recently schemas.xmlsoap.org website
configured HTTP to HTTPS redirect and in doing so broke our WSDL parser
as it will try to fetch without following HTTP redirects. This results
in empty HTTP body that the parser tries to parse and fails with:

```
org.xml.sax.SAXParseException: Premature end of file.
```

To resolve this this stores the two identified XSD files locally and
sets the custom EntityResolver via a custom WSDLLocator to load those
instead. The same is done for the CXF schema parsing classes which can
load the catalog files from `META-INF/jax-ws-catalog.xml`.