fix(deps): force declared Spring Framework version
We can get different version of Spring Framework via Camel dependencies,
e.g. via `camel-spring`, even though we wish to be aligned with Camel
there are situations we might want a newer version (but compatible) of
Spring. This forces the declared version of Spring Framework by
importing the spring-framework-bom with the same version before the
Camel BOM, and that way solidifying the version of Spring Framework to
be used.