chore(meta): integration tests
Given that we don't have any integration tests we can't be alerted to
failures of the syndesis-meta configuration. In this particular case we
were not aware that a MongoDB auto-configuration was enabled to be
loaded as the MongoDB classes are on the classpath but failed to load as
a dependency of theirs was not present on the classpath.

```
java.lang.IllegalStateException: Failed to introspect Class [org.springframework.boot.actuate.autoconfigure.metrics.mongo.MongoMetricsAutoConfiguration$MongoConnectionPoolMetricsConfiguration] from ClassLoader [org.springframework.boot
.loader.LaunchedURLClassLoader@4459eb14]

ClassNotFoundException: io.micrometer.core.instrument.binder.mongodb.MongoConnectionPoolTagsProvider
```